reduce to 1 worker
