+++ nala/libs/auth.setup.cjs
@@ -1,6 +1,7 @@
 /* eslint-disable import/no-import-module-exports */
 import { test as setup, expect } from '@playwright/test';
 import path from 'path';
+import { installEdsThrottleOnPage } from './eds-throttle.js';
 
 const authFile = path.join(__dirname, '../../nala/.auth/user.json');
 
@@ -14,6 +15,7 @@ setup('authenticate, @mas-studio', async ({ page, baseURL, browserName }) => {
     expect(process.env.IMS_EMAIL, 'ERROR: No environment variable for email provided for IMS Test.').toBeTruthy();
     expect(process.env.IMS_PASS, 'ERROR: No environment variable for password provided for IMS Test.').toBeTruthy();
 
+    await installEdsThrottleOnPage(page);
     await page.goto(`${baseURL}/studio.html`);
     await page.waitForURL('**/auth.services.adobe.com/en_US/index.html**/');
