+++ nala/libs/global-request-counter.js
@@ -5,6 +5,7 @@
 
 import * as fs from 'fs';
 import * as path from 'path';
+import { resolveEdsMaxRps, isEdsEdgeHost, throttleEdsGap, logEdsThrottleOnce } from './eds-throttle.js';
 
 const DEFAULT_TRACKED_URLS = {
     ODIN_AEM: 'https://author-p22655-e59433.adobeaemcloud.com',
@@ -32,54 +33,6 @@ if (!globalThis.requestCounter) {
     };
 }
 
-/** EDS / Helix edge (~200 rps cap). Lower workers does not cap in-flight requests per page. */
-function resolveEdsMaxRps() {
-    if (process.env.NALA_EDS_THROTTLE_DISABLED === '1') return 0;
-    if (process.env.NALA_EDS_MAX_RPS !== undefined && process.env.NALA_EDS_MAX_RPS !== '') {
-        const v = Number.parseInt(process.env.NALA_EDS_MAX_RPS, 10);
-        return Number.isFinite(v) && v > 0 ? v : 0;
-    }
-    return process.env.CI === 'true' ? 150 : 0;
-}
-
-function isEdsEdgeHost(url) {
-    try {
-        const { hostname } = new URL(url);
-        return (
-            hostname.endsWith('.aem.live') ||
-            hostname.endsWith('.hlx.page') ||
-            hostname.endsWith('.hlx.live') ||
-            hostname === 'aem.live'
-        );
-    } catch {
-        return false;
-    }
-}
-
-/**
- * Serialize route.continue() for EDS hosts with a minimum gap so burst RPS stays bounded.
- * Shared per worker (globalThis).
- */
-function throttleEdsGap(maxRps) {
-    const minGapMs = 1000 / maxRps;
-    if (!globalThis._edsThrottleChain) {
-        globalThis._edsThrottleChain = Promise.resolve();
-    }
-
-    const next = globalThis._edsThrottleChain.then(async () => {
-        const last = globalThis._edsThrottleLastContinueAt ?? 0;
-        const now = Date.now();
-        const wait = Math.max(0, minGapMs - (now - last));
-        if (wait > 0) {
-            await new Promise((r) => setTimeout(r, wait));
-        }
-        globalThis._edsThrottleLastContinueAt = Date.now();
-    });
-
-    globalThis._edsThrottleChain = next.catch(() => {});
-    return next;
-}
-
 class GlobalRequestCounter {
     /**
      * Initialize global request tracking for multiple services
@@ -98,12 +51,7 @@ class GlobalRequestCounter {
         }
 
         const edsMaxRps = resolveEdsMaxRps();
-        if (edsMaxRps > 0 && !globalThis._edsThrottleLogged) {
-            console.info(
-                `[NALA] EDS request pacing enabled (~${edsMaxRps} rps max for .aem.live / hlx hosts). Set NALA_EDS_THROTTLE_DISABLED=1 to turn off.\n`,
-            );
-            globalThis._edsThrottleLogged = true;
-        }
+        logEdsThrottleOnce(edsMaxRps);
 
         // Set up routing to track requests to all configured URLs
         await page.route('**/*', async (route) => {
