+++ playwright.config.js
@@ -29,14 +29,14 @@ const config = {
     /* Retry on CI only */
     retries: process.env.CI ? 1 : 0,
     /*
-     * EDS (aem.live) rate-limits around ~200 RPS per edge context. Each worker drives a full
-     * browser; each page load fans out to many parallel requests. Too many workers → 429s and
-     * flaky "app not loading" in CI. Override with NALA_PLAYWRIGHT_WORKERS (e.g. 1) if needed.
+     * EDS (~200 rps / hostname). EDS pacing in nala/libs/eds-throttle.js is per *worker*; multiple
+     * workers multiply traffic to the same preview host → 429s. Default CI workers=1; override
+     * with NALA_PLAYWRIGHT_WORKERS only if EDS grants a higher automation cap.
      */
     workers: (() => {
         const fromEnv = Number.parseInt(process.env.NALA_PLAYWRIGHT_WORKERS ?? '', 10);
         if (Number.isFinite(fromEnv) && fromEnv > 0) return fromEnv;
-        return process.env.CI ? 2 : 3;
+        return process.env.CI ? 1 : 3;
     })(),
     /* Reporter to use. */
     reporter: process.env.CI
